requirements.txt
# MISIS FAQ Bot Dependencies

aiogram>=3.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pymorphy2>=0.9.1
pymorphy2-dicts-ru>=2.4.0
